+++ internal/commands/database/index/delete.go
@@ -9,6 +9,7 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
+	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -36,7 +37,8 @@ func (s *deleteCommand) InitCommand() {
 	flagSet.StringVar(&s.name, "name", "", "Index name")
 	s.AddFlags(flagSet)
 
-	_ = s.Cobra().MarkFlagRequired("name")
+	commands.Must(s.Cobra().MarkFlagRequired("name"))
+	commands.Must(s.Cobra().RegisterFlagCompletionFunc("name", cobra.NoFileCompletions))
 }
 
 // Execute implements commands.MultipleArgumentCommand