+++ internal/commands/kubernetes/create.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
+	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -147,9 +148,10 @@ func (c *createCommand) InitCommand() {
 	config.AddToggleFlag(fs, &c.params.wait, "wait", false, "Wait for cluster to be in running state before returning.")
 	c.AddFlags(fs)
 
-	_ = c.Cobra().MarkFlagRequired("name")
-	_ = c.Cobra().MarkFlagRequired("network")
-	_ = c.Cobra().MarkFlagRequired("zone")
+	commands.Must(c.Cobra().MarkFlagRequired("name"))
+	commands.Must(c.Cobra().MarkFlagRequired("network"))
+	commands.Must(c.Cobra().MarkFlagRequired("zone"))
+	commands.Must(c.Cobra().RegisterFlagCompletionFunc("name", cobra.NoFileCompletions))
 
 	// Deprecating k8s
 	// TODO: Remove this in the future
@@ -157,9 +159,9 @@ func (c *createCommand) InitCommand() {
 }
 
 func (c *createCommand) InitCommandWithConfig(cfg *config.Config) {
-	_ = c.Cobra().RegisterFlagCompletionFunc("network", namedargs.CompletionFunc(completion.Network{}, cfg))
-	_ = c.Cobra().RegisterFlagCompletionFunc("zone", namedargs.CompletionFunc(completion.Zone{}, cfg))
-	_ = c.Cobra().RegisterFlagCompletionFunc("version", namedargs.CompletionFunc(completion.KubernetesVersion{}, cfg))
+	commands.Must(c.Cobra().RegisterFlagCompletionFunc("network", namedargs.CompletionFunc(completion.Network{}, cfg)))
+	commands.Must(c.Cobra().RegisterFlagCompletionFunc("zone", namedargs.CompletionFunc(completion.Zone{}, cfg)))
+	commands.Must(c.Cobra().RegisterFlagCompletionFunc("version", namedargs.CompletionFunc(completion.KubernetesVersion{}, cfg)))
 }
 
 // ExecuteWithoutArguments implements commands.NoArgumentCommand