+++ internal/commands/kubernetes/nodegroup/delete.go
@@ -8,6 +8,7 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
+	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -35,7 +36,8 @@ func (s *deleteCommand) InitCommand() {
 	flagSet.StringVar(&s.name, "name", "", "Node group name")
 	s.AddFlags(flagSet)
 
-	_ = s.Cobra().MarkFlagRequired("name")
+	commands.Must(s.Cobra().MarkFlagRequired("name"))
+	commands.Must(s.Cobra().RegisterFlagCompletionFunc("name", cobra.NoFileCompletions))
 }
 
 // ExecuteSingleArgument implements commands.SingleArgumentCommand