+++ internal/commands/server/networkinterface/delete.go
@@ -8,6 +8,7 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
+	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -36,7 +37,8 @@ func (s *deleteCommand) InitCommand() {
 	fs.IntVar(&s.interfaceIndex, "index", 0, "Interface index.")
 
 	s.AddFlags(fs)
-	_ = s.Cobra().MarkFlagRequired("index")
+	commands.Must(s.Cobra().MarkFlagRequired("index"))
+	commands.Must(s.Cobra().RegisterFlagCompletionFunc("index", cobra.NoFileCompletions))
 }
 
 // MaximumExecutions implements Command.MaximumExecutions